moved cli to src/rail/cli/rail_pipe
