+++ src/rail/cli/rail_pipe/__main__.py
@@ -1,0 +1,5 @@
+# This file must exist with these contents
+from .pipe_commands import pipe_cli
+
+if __name__ == "__main__":
+    pipe_cli()
