+++ src/rail/cli/rail_pipe/reduce_roman_rubin_data.py
@@ -7,7 +7,7 @@ import pyarrow.dataset as ds
 import pyarrow.parquet as pq
 from pyarrow import acero
 
-from rail.cli.pipe_options import RunMode
+from rail.cli.rail_pipe.pipe_options import RunMode
 from rail.utils.project import RailProject
